nmtCov: recheck the file still exist before trying to remove them

+++ tjpcov/nmt_tools.py
@@ -194,8 +194,10 @@ def get_workspace(f1, f2, m1, m2, bins, outdir, **kwargs):
         # Recheck that the file has not been written by other proccess
         if fname and not os.path.isfile(fname):
             w.write_to(fname)
-        if isfile2:
-            # Remove the other to avoid later confusions
+        # Check if the other files exist. Recheck in case other process has
+        # removed it in the mean time.
+        if isfile2 and os.path.isfile(fname2):
+            # Remove the other to avoid later confusions.
             os.remove(fname2)
     elif isfile:
         w.read_from(fname)
@@ -259,7 +261,9 @@ def get_covariance_workspace(f1, f2, f3, f4, m1, m2, m3, m4, outdir, **kwargs):
             cw.write_to(fnames[0])
         for fn, isf in zip(fnames[1:], isfiles[1:]):
             # This will only be run if they don't exist or recompute = True
-            if isf:
+            # Recheck the file exist in case other process has removed it in
+            # the mean time.
+            if isf and os.path.isfile(fn):
                 # Remove old covariance workspace if you have recomputed it
                 os.remove(fn)
     else:
